line 184 type error fixed

+++ train.py
@@ -181,7 +181,7 @@ def validate(val_list, model, criterion):
         img = Variable(img)
         output = model(img)
         
-        mae += abs(output.data.sum()-target.sum().cuda())
+        mae += abs(output.data.sum()-target.sum().type(torch.DoubleTensor).cuda())
         
     mae = mae/len(test_loader)    
     print(' * MAE {mae:.3f} '
